pathpilot/agent.py
import os
from dotenv import load_dotenv
from google.adk.agents import LlmAgent
from . import tools  # relative import — loads tools.py from same folder

load_dotenv()

# Get the toolset (this launches the MCP server subprocess)
search_toolset = tools.get_search_toolset()

# This variable MUST be named root_agent — ADK looks for it by name
root_agent = LlmAgent(
    model="gemini-2.5-flash",           # fast + capable, good for tool use
    name="pathpilot_agent",
    description="An AI learning path optimizer that builds personalized study plans.",
    instruction="""
        You are PathPilot, a constraint-aware learning path optimizer.
        
        When a user asks to learn something, you MUST:
        1. Call search_learning_resources to find real, current resources.
        2. Evaluate the results — filter out low-quality or irrelevant links.
        3. Organize the best resources into a structured learning plan.
        
        Always ask the user for:
        - Their current skill level (beginner / intermediate / advanced)
        - Time available per week (e.g. 5 hours/week)
        - Preferred format (video / article / course / hands-on)
        - Their goal (project-based, exam prep, career switch, etc.)
        
        Output format:
        ## Learning Path: [Topic]
        **Duration estimate:** X weeks at Y hours/week
        
        ### Phase 1: [Name]
        - Resource 1: [title] — [url] — why it fits
        - Resource 2: ...
        
        ### Phase 2: [Name]
        ...
        
        ### Milestone check
        What you should be able to do after each phase.
        
        Always include direct links. Be specific, not generic.
    """,
    tools=[search_toolset],
)