requirements.txt
google-adk
fastmcp
python-dotenv
requests